+++ scout/server/blueprints/institutes/templates/overview/institute_sidebar.html
@@ -3,15 +3,8 @@
   <div id="sidebar-container" class="sidebar-expanded d-none d-md-block"><!-- d-* hiddens the Sidebar in smaller devices. Its itens can be kept on the Navbar 'Menu' -->
       <!-- Bootstrap List Group -->
       <ul class="list-group">
-          <a href="#" data-toggle="sidebar-collapse" class="bg-dark list-group-item list-group-item-action d-flex align-items-center">
-              <div class="d-flex w-100 justify-content-start align-items-center">
-                  <span id="collapse-icon" class="fa fa-2x mr-3"></span>
-                  <span id="collapse-text" class="menu-collapsed">Collapse</span>
-              </div>
-          </a>
-
           <!-- cases menu item -->
-          <a href="{{ url_for('cases.index', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start">
+          <a href="{{ url_for('overview.cases', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start">
               <div class="d-flex w-100 justify-content-start align-items-center">
                   <span class="fa fa-sitemap  mr-3"></span>
                   <span class="menu-collapsed">Cases</span>
@@ -47,8 +40,10 @@
               <div class="d-flex w-100 justify-content-start align-items-center">
                   <span class="fa fa-share-alt mr-3"></span>
                   <span class="menu-collapsed">Shared data</span>
+                  <span class="submenu-icon ml-auto"></span>
               </div>
           </a>
+        {{ shared_submenu(institute) }}
 
           <!-- general settings menu item -->
           <a href="{{ url_for('overview.institute_settings', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start">
@@ -59,4 +54,22 @@
           </a>
       </ul><!-- List Group END-->
   </div><!-- sidebar-container END -->
+{% endmacro %}
+
+
+{% macro shared_submenu(institute) %}
+<!-- Submenu content -->
+<div id='submenu1' class="collapse sidebar-submenu">
+    <div href="#" class="bg-dark list-group-item text-white">
+      <div class="flex-fill bd-highlight">
+        <a href="{{ url_for('overview.clinvar_submissions', institute_id=institute._id) }}"><span class="menu-collapsed ml-2 align-left">ClinVar</span></a>
+      </div>
+    </div>
+
+    <div href="#" class="bg-dark list-group-item text-white">
+        <div class="flex-fill bd-highlight disabled">
+          <a href="#" target="_blank"><span class="menu-collapsed ml-2 align-left">MatchMaker</span></a>
+        </div>
+    </div>
+</div>
 {% endmacro %}
