+++ scout/server/blueprints/institutes/views.py
@@ -65,6 +65,62 @@ def institutes():
     return render_template("overview/institutes.html", **data)
 
 
+@blueprint.route("/<institute_id>/cases")
+@templated("overview/cases.html")
+def cases(institute_id):
+    """Display a list of cases for an institute."""
+
+    institute_obj = institute_and_case(store, institute_id)
+    query = request.args.get("query")
+
+    limit = 100
+    if request.args.get("limit"):
+        limit = int(request.args.get("limit"))
+
+    skip_assigned = request.args.get("skip_assigned")
+    is_research = request.args.get("is_research")
+    all_cases = store.cases(
+        collaborator=institute_id,
+        name_query=query,
+        skip_assigned=skip_assigned,
+        is_research=is_research,
+    )
+
+    sort_by = request.args.get("sort")
+    sort_order = request.args.get("order") or "asc"
+    if sort_by:
+        pymongo_sort = pymongo.ASCENDING
+        if sort_order == "desc":
+            pymongo_sort = pymongo.DESCENDING
+        if sort_by == "analysis_date":
+            all_cases.sort("analysis_date", pymongo_sort)
+        elif sort_by == "track":
+            all_cases.sort("track", pymongo_sort)
+        elif sort_by == "status":
+            all_cases.sort("status", pymongo_sort)
+
+    LOG.debug("Prepare all cases")
+
+    prioritized_cases = store.prioritized_cases(institute_id=institute_id)
+
+    data = controllers.cases(store, all_cases, prioritized_cases, limit)
+    data["sort_order"] = sort_order
+    data["sort_by"] = sort_by
+    data["nr_cases"] = store.nr_cases(institute_id=institute_id)
+
+    sanger_unevaluated = controllers.get_sanger_unevaluated(store, institute_id, current_user.email)
+    if len(sanger_unevaluated) > 0:
+        data["sanger_unevaluated"] = sanger_unevaluated
+
+    return dict(
+        institute=institute_obj,
+        skip_assigned=skip_assigned,
+        is_research=is_research,
+        query=query,
+        **data,
+    )
+
+
 @blueprint.route("/<institute_id>/causatives")
 @templated("overview/causatives.html")
 def causatives(institute_id):
