+++ scout/server/blueprints/variant/templates/variant/clinvar.html
@@ -22,7 +22,7 @@
 
 {% block top_nav %}
   <li class="nav-item">
-    <a class="nav-link" href="{{ url_for('cases.cases', institute_id=institute._id) }}">
+    <a class="nav-link" href="{{ url_for('overview.cases', institute_id=institute._id) }}">
       {{ institute.display_name }}
     </a>
   </li>
