+++ scout/server/blueprints/variants/templates/variants/variants.html
@@ -17,7 +17,7 @@
     <a class="nav-link text-nowrap" href="{{ url_for('cases.index') }}">Institutes</a>
   </li>
   <li class="nav-item">
-    <a class="nav-link text-nowrap" href="{{ url_for('cases.cases', institute_id=institute._id) }}">
+    <a class="nav-link text-nowrap" href="{{ url_for('overview.cases', institute_id=institute._id) }}">
       {{ institute.display_name }}
     </a>
   </li>
@@ -178,7 +178,7 @@
 
     var cytoStart = document.getElementById("cytoband_start");
     initSearchConstraints(cytoStart, startElem)
-    
+
     var cytoEnd = document.getElementById("cytoband_end");
     initSearchConstraints(cytoEnd, endElem)
 
@@ -187,7 +187,7 @@
     window.onload=function() {
       populateCytobands({{cytobands|safe}});
     }
-    
+
     $('select[multiple]').selectpicker({
       width: '100%'
     });
